refactor: factory 클래스 제거 Car객체가 다형성이 적용되는 객체가 아니기때문에 사용 의미가 없다고 판단
