+++ src/main/java/racewinner/controller/RaceApp.java
@@ -2,8 +2,6 @@ package racewinner.controller;
 
 
 import racewinner.domain.CarRacing;
-import racewinner.domain.Cars;
-import racewinner.domain.CarsFactory;
 import racewinner.util.RandomGenerator;
 import racewinner.view.InputView;
 import racewinner.view.ResultView;
@@ -13,8 +11,7 @@ public class RaceApp {
         final String carNameList = InputView.inputCarName();
         final int raceAttemptCount = InputView.inputRaceAttemptCount();
 
-        final Cars cars = new CarsFactory().create(carNameList);
-        final CarRacing carRacing = new CarRacing(cars, raceAttemptCount, new RandomGenerator());
+        final CarRacing carRacing = new CarRacing(carNameList, raceAttemptCount, new RandomGenerator());
 
         while (!carRacing.isFinished()) {
             carRacing.race();
